Final FIxes

+++ Symfony Self Care Application/src/Helper/FileUploader.php
@@ -5,6 +5,7 @@ namespace App\Helper;
 use Aws\S3\S3Client;
 use Aws\S3\Exception\S3Exception;
 use Exception;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
@@ -13,16 +14,44 @@ use Symfony\Component\String\Slugger\SluggerInterface;
 
 class FileUploader
 {
+    /** @var SluggerInterface */
     private SluggerInterface $slugger;
+
+    /** @var KernelInterface */
     private KernelInterface $kernel;
+
+    /** @var S3Client */
     private S3Client $s3Client;
+
+    /** @var string */
     private string $bucket;
 
-    public function __construct(SluggerInterface $slugger, KernelInterface $kernel, string $accessKey, string $secretKey, string $region, string $endpoint, string $bucket)
+    /** @var DefaultHelper */
+    protected DefaultHelper $helper;
+
+    /**
+     * @var LoggerInterface
+     */
+    protected LoggerInterface $fileUploadRemoveLogger;
+
+    /**
+     * @param SluggerInterface $slugger
+     * @param KernelInterface $kernel
+     * @param DefaultHelper $helper
+     * @param LoggerInterface $fileUploadRemoveLogger
+     * @param string $accessKey
+     * @param string $secretKey
+     * @param string $region
+     * @param string $endpoint
+     * @param string $bucket
+     */
+    public function __construct(SluggerInterface $slugger, KernelInterface $kernel, DefaultHelper $helper, LoggerInterface $fileUploadRemoveLogger, string $accessKey, string $secretKey, string $region, string $endpoint, string $bucket)
     {
         $this->slugger = $slugger;
         $this->kernel = $kernel;
         $this->bucket = $bucket;
+        $this->helper = $helper;
+        $this->fileUploadRemoveLogger = $fileUploadRemoveLogger;
 
         $this->s3Client = new S3Client([
             'version' => 'latest',
@@ -80,6 +109,7 @@ class FileUploader
             try {
                 $fileName = $this->upload($file, $folderPath, $renameFile);
             } catch (FileException $e) {
+                $this->fileUploadRemoveLogger->error($e->getMessage(), ['fileName' => $fileName]);
                 $success = false;
             }
         } else {
@@ -106,13 +136,14 @@ class FileUploader
     {
         $status = true;
 
-        if (!empty($fileName)) {
+        if (!empty($fileName) && $fileName !== $this->helper->getEnvValue('app_default_image')) {
             try {
                 $this->s3Client->deleteObject([
                     'Bucket' => $this->bucket,
                     'Key' => $folderPath . $fileName,
                 ]);
             } catch (S3Exception $e) {
+                $this->fileUploadRemoveLogger->error($e->getMessage(), ['fileName' => $fileName]);
                 $status = false;
             }
         }
